This commit adds a very initial support for Hassos + visual interface (Cog via DRM or Weston) on the top of the ComfilePi B070WR.
The Comfile is basically a Raspberry CM3+ driving very known hardware, most of it already sold by Adrafruit or other dealers.

For further information, see http://comfilewiki.co.kr/en/doku.php?id=comfilepi:index

The present status is that the browser (from now on COG) can be launched but it is useless since the touchscreen is not recognised. Also a onscreen keyboard is missing.

This must be refined. At this stage, version of COG is 0.8.0. As we can see, cog is moving rapidly to 0.10 and many of this issues seem to be solved.

In the TODO list, although not exhaustive:
* Launch COG as service or as a docker
* Investigate the usage of present version of COG or pull ahead development version 0.9.2
* Prepare support for the i2c buses in modules.conf
* Prepare integration in Home Assistant with expansion board ComfilePi CP-IO22
* Support buzzer with touchscreen
* Clean unnecesary packages; most probably Weston will be dropped in favour of DRM